Improve the rockbuilder build logic and config options (#795)
Updated rockbuilder logic and config options:

- update the default rocm_home default seach path after the code was
moved to another directory: ../../dist/build/rocm
- ROCM_HOME env variable can be used to specify a different location for
rocm-sdk to be used
- use patches from externa-builds/pytorch/patches folder instead of
forking it
- support building pytorch, pytorch audio and vision also on windows
without requiring a bash prompt
- add option to specify environment variables
- environment variables are set at the time when project phases are
executed and then reset back to previous values or removed after the
project has been build
      (env_common, env_linux, env_windows)
 - git am fix for hipify run on windows
- Hipified patches could fail to apply with command "git am" if the
hipify command was executed on dos-prompt instead of using a
bash-prompt.
- Reason for the failure was that the original hipify patches expected
that they would be applied to files having Linux line ending. When
hipify is executed on dos-prompt they wil however have a dos line ending
and that caused the patched to file.
    - This is fixed with '--ignore-whitespace' flag for git am.
- Update messages printed out during the build steps
- Improve code checkout and hipify phases
- Add separate/optional hipify_cmd to config files instead of using the
config-command to execute the hipify because we need to do in the
background do additional things during the hipify phase
          a. If hipify command is specified in config file
- run hipify command to scan and convert files to hip/rocm format
            - commit code changes made by hipify command to own commit
          b. If patches/hipified directory exist
- apply patches from that directory even if the hipify_cmd is not
specified
- Add optional pre_config and post_config command phases
- Perform "git reset --hard" automatically before trying to execute the
git checkout command.
- This fixes git checkout failures in cases where the project directory
contains modifications to files that have not been committed. This can
happen for the files for example if the previously executed build
command has modified them
- Add error checks, error message printouts and process termination if
the command execution failed
    - added support for commands having multiple lines
        - all commands executed can now contain from more than
          one line (earlier restricted for the build command)
- Added support for specifying cmake config commands
- If cmake config options are specified, we will execue cmake-config,
build and install steps automatically in addition of executing other
config/build/install commands that can also be present in the
config-file. For example the amd-smi project build and install will
rquire the usage of both the cmake specific configure, build and install
steps and then the execution of 'pip install'-command
- Added more projects to test the building

---------

Signed-off-by: Mika Laitio <lamikr@gmail.com>
Signed-off-by: Mika Laitio <mika.laitio@amd.com>