initial support for building therock itself
rockbuilder is now a separate project from therock and
we can not anymore expect that therock has been build
to directory ../../build/dist/rocm

Instead we should offer the option to checkout therock
source code and then use that to build therock sdk in
directory src_projects/therock.

Initial support for that is now implemented by adding
a therock.cfg project to build it.

Integration to rockbuilder_cfg works for selecting the target GPU
but therock build itself must still be done just like with any
other apps build by the rockbuilder.
(Except the addition of SKIP_ROCM_HOME_CHECK=1 flag that is planned
to be removed in future once things gets more stable)

TheRock can be build with commands:

  $ source ./init_rcb_env.sh
  $ SKIP_ROCM_HOME_CHECK=1 ./rockbuilder.py --project projects/therock.cfg

Then rest of the apps can be build just like earlier with commands:

  $ export ROCM_HOME=src_projects/therock/build/dist/rocm
  $ ./rockbuilder.py --project ./projects/pytorch_28_amd_apps.pcfg

Signed-off-by: Mika Laitio <mika.laitio@amd.com>