fix the build phase stamp cleanup
./rockbuilder.py --build command
cleaned up also all command phase stamps that
are executed before the build command.

This happened because the init-phase that is the
first phase, is executed always.

Fix so that the stamp files for the phases after
init are only cleaned if the "--init" phase was
given explicitly as a parameter.

Signed-off-by: Mika Laitio <lamikr@gmail.com>