aotriton support MAX_JOBS to limit build cpu-count
Add support for the MAX_JOBS environment variable
to limit the amount of cpu cores used for building the
aotriton and all of it's kernels.

This helps to get aotriton build on systems having less
than optimal amount of memory without processes gettting
killed. (32GB on 8 core/16 thread system for example)

Signed-off-by: Mika Laitio <lamikr@gmail.com>