add sglang_053.apps to test it
Add initial sglang 0.5.3 version to test the building
and usage. This version will most likely still have issues
that needs to be fixed but because it adds also new apps it's recipe
is good to be merged. New apps that 0.5.3 requires compared to 0.5.0
are:

apps/fast-hadamard-transform.cfg
apps/mooncake.cfg
apps/yalantinglibs.cfg

Signed-off-by: Mika Laitio <mika.laitio@amd.com>